Address code review issues
